feat: Implement ClientProjects component to display project listings with loading state
